Pause breaks with minified KO
When a minified KO is used, the underlying calls to "valueWillMutate" and "valueHasMutated" are not called using an array notation (e.g. observable["valueWillMutate"]). This results in certain array calls, such as "remove", to fail when this extension is used. To work around this we make sure to add a reference to the functions using the minified name.